test 1

+++ src/include/board/k_gate_module_v3.h
@@ -94,7 +94,7 @@ uint8 supla_board_gate(void);
 											supla_log(LOG_DEBUG, "IP FIELD = %i", ipaddr_addr(SUPLA_ESP_SOFTVER)); };	\
 										if ( ChannelNumber == 0 ) {	\
 											int gate_status;	\
-											gate_status = supla_board_gate;	\
+											gate_status = (int)supla_board_gate;	\
 											state->Fields |= SUPLA_CHANNELSTATE_FIELD_LASTCONNECTIONRESETCAUSE;	\
 											state->LastConnectionResetCause = gate_status; };	\
